feat: SignalBroker is now exposing the observed signals.
feat: SignalBroker emits a signal for every added signal. This is needed for extending signal functionality.

feat: SignalBroker is now exposing a task that executes all stacks and clears them. This is needed in some circumstances for signals that are not executed and therefore queued.

refactor: Removed all Meta classes for simplicity. Now the Task class has parameters for setting isMeta and isUnique.